feat: make experimental default and improve test cmd
