+++ package.json
@@ -31,13 +31,12 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@babel/cli": "7.1.2",
-    "@babel/core": "7.1.2",
-    "@babel/plugin-transform-regenerator": "7.0.0",
-    "@babel/plugin-transform-runtime": "7.1.0",
-    "@babel/preset-env": "7.1.0",
+    "@babel/cli": "7.2.0",
+    "@babel/core": "7.2.0",
+    "@babel/plugin-transform-runtime": "7.2.0",
+    "@babel/preset-env": "7.2.0",
     "@babel/register": "7.0.0",
-    "@babel/runtime": "7.1.2",
+    "@babel/runtime": "7.2.0",
     "@commitlint/config-conventional": "^7.1.2",
     "@commitlint/lint": "^7.2.1",
     "@commitlint/load": "^7.2.1",
@@ -46,13 +45,11 @@
     "async": "^2.6.1",
     "babel-loader": "8.0.4",
     "babel-plugin-transform-flow-comments": "^6.22.0",
-    "browserify-zlib": "~0.2.0",
     "chalk": "^2.4.1",
     "clean-documentation-theme": "~0.5.2",
     "codecov": "^3.1.0",
     "conventional-changelog": "^2.0.3",
     "conventional-github-releaser": "^2.0.0",
-    "del": "^3.0.0",
     "detect-node": "^2.0.4",
     "documentation": "^9.0.0-alpha.1",
     "es6-promisify": "^6.0.1",
@@ -64,54 +61,46 @@
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
     "execa": "^1.0.0",
-    "filesize": "^3.6.1",
+    "execa-output": "^2.0.0",
     "findup-sync": "^2.0.0",
     "fs-extra": "^7.0.0",
     "gh-pages": "^2.0.1",
     "git-validate": "^2.2.4",
     "globby": "^8.0.1",
     "joi": "^14.0.1",
-    "json-loader": "~0.5.7",
-    "karma": "^3.1.1",
+    "karma": "3.1.1",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
+    "karma-cli": "^2.0.0",
     "karma-edge-launcher": "~0.4.2",
     "karma-firefox-launcher": "^1.1.0",
     "karma-junit-reporter": "^1.2.0",
     "karma-mocha": "^1.3.0",
-    "karma-mocha-own-reporter": "git+https://github.com/dryajov/karma-mocha-own-reporter#d562a92a12d5c76469a05d67cee19bcb8db22b23",
+    "karma-mocha-reporter": "^2.2.5",
     "karma-mocha-webworker": "^1.3.0",
     "karma-sourcemap-loader": "~0.3.7",
-    "karma-webpack": "v4.0.0-beta.0",
+    "karma-webpack": "4.0.0-beta.0",
     "listr": "~0.14.2",
-    "listr-verbose-renderer": "~0.4.1",
+    "listr-verbose-renderer": "^0.5.0",
     "lodash": "^4.17.11",
     "mocha": "^5.2.0",
     "mocha-jenkins-reporter": "~0.4.1",
-    "npm-which": "^3.0.1",
     "nyc": "^13.1.0",
-    "p-map": "^2.0.0",
     "pify": "^4.0.1",
-    "pretty-hrtime": "^1.0.3",
     "prompt-promise": "^1.0.3",
     "read-pkg-up": "^4.0.0",
     "resolve-bin": "~0.4.0",
     "rimraf": "^2.6.2",
+    "rxjs": "^6.3.3",
     "semver": "^5.6.0",
     "simple-git": "^1.105.0",
     "stats-webpack-plugin": "~0.7.0",
-    "stream-http": "^3.0.0",
     "terser-webpack-plugin": "^1.1.0",
-    "through": "^2.3.8",
-    "transform-loader": "~0.2.4",
-    "uglify-es": "^3.3.9",
     "update-notifier": "^2.5.0",
     "webpack": "^4.23.1",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.1.2",
     "webpack-merge": "^4.1.4",
-    "yargs": "^12.0.2",
-    "yargs-parser": "^11.0.0"
+    "yargs": "^12.0.2"
   },
   "repository": {
     "type": "git",
