+++ src/config/karma.conf.js
@@ -1,65 +1,88 @@
 'use strict'
 
-let concurrency = 1
-let reporters = ['mocha-own']
-
-if (process.env.CI) {
-  reporters.push('junit')
+const merge = require('webpack-merge')
+const webpack = require('webpack')
+const webpackConfig = require('./webpack.config')
+const { fromRoot, hasFile } = require('../utils')
+const isWebworker = process.env.AEGIR_WEBWORKER === 'true'
+const isProduction = process.env.NODE_ENV === 'production'
+const userConfig = require('./user')()
+const env = {
+  TEST_DIR: JSON.stringify(fromRoot('test')),
+  TEST_BROWSER_JS: hasFile('test', 'browser.js')
+    ? JSON.stringify(fromRoot('test', 'browser.js'))
+    : JSON.stringify('')
 }
+const karmaWebpackConfig = merge(webpackConfig({ production: isProduction }), {
+  entry: '',
+  devtool: 'inline-source-map',
+  output: {
+    libraryTarget: 'var'
+  },
+  plugins: [
+    new webpack.DefinePlugin(env)
+  ]
+})
 
-let browsers = []
+const karmaConfig = (config) => {
+  return {
+    browsers: ['ChromeHeadless'],
+    frameworks: isWebworker ? ['mocha-webworker'] : ['mocha'],
+    basePath: process.cwd(),
+    files: [
+      {
+        pattern: 'node_modules/aegir/src/config/karma-entry.js',
+        watched: false,
+        included: !isWebworker
+      },
+      {
+        pattern: 'test/fixtures/**/*',
+        watched: false,
+        served: true,
+        included: false
+      }
+    ],
 
-if (process.env.TRAVIS) {
-  browsers.push('FirefoxCustom')
-} else {
-  browsers.push('ChromeCustom')
-}
+    preprocessors: {
+      'node_modules/aegir/src/config/karma-entry.js': [ 'webpack', 'sourcemap' ]
+    },
 
-const browsersEnv = process.env.AEGIR_BROWSERS
-if (browsersEnv) {
-  if (browsersEnv.indexOf(',') !== -1) {
-    browsers = browsersEnv.split(',')
-  } else {
-    browsers = [browsersEnv]
+    client: {
+      mochaWebWorker: {
+        pattern: [
+          'node_modules/aegir/src/config/karma-entry.js'
+        ]
+      }
+    },
+
+    webpack: karmaWebpackConfig,
+
+    webpackMiddleware: {
+      stats: 'errors-only'
+    },
+
+    reporters: [
+      'mocha'
+    ],
+
+    mochaReporter: {
+      output: 'autowatch'
+    },
+
+    junitReporter: {
+      outputDir: process.cwd(),
+      outputFile: isWebworker ? 'junit-report-webworker.xml' : 'junit-report-browser.xml',
+      useBrowserName: false
+    },
+
+    logLevel: config.LOG_ERROR,
+    autoWatch: false,
+    singleRun: true,
+    colors: true,
+    browserNoActivityTimeout: 50 * 1000
   }
 }
 
-module.exports = function (config) {
-  const randomNumber = Math.floor(Math.random() * 10000)
-  const junitFile = `junit-report-browser-${randomNumber}.xml`
-
-  config.set({
-    frameworks: ['mocha'],
-    basePath: process.cwd(),
-    webpackMiddleware: {
-      noInfo: true
-    },
-    plugins: config.plugins.concat([
-      'karma-webpack',
-      'karma-sourcemap-loader',
-      'karma-mocha',
-      'karma-mocha-webworker',
-      'karma-mocha-own-reporter',
-      'karma-junit-reporter',
-      'karma-chrome-launcher',
-      'karma-firefox-launcher'
-    ].map(m => require(m))),
-    reporters: reporters,
-    mochaOwnReporter: {
-      reporter: 'spec'
-    },
-    junitReporter: {
-      outputDir: process.cwd(),
-      outputFile: junitFile,
-      useBrowserName: false
-    },
-    port: 9876,
-    colors: true,
-    logLevel: config.LOG_WARN,
-    autoWatch: false,
-    browsers: browsers,
-    singleRun: true,
-    concurrency: concurrency,
-    failOnEmptyTestSuite: true
-  })
+module.exports = (config) => {
+  config.set(merge(karmaConfig(config), userConfig.karma))
 }
